pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "observable-llm-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "pydantic",
  "python-dotenv",
  "requests"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
